+++ paddle/phi/kernels/gpu/gather_kernel.cu
@@ -30,7 +30,7 @@ void GatherKernel(const Context& dev_ctx,
   const auto& index_type = index.dtype();
   auto axis_v = axis.to<int>();
   if (axis_v < 0) {
-    axis_v += x.dims().size();
+    axis_v += static_cast<int>(x.dims().size());
   }
   if (axis_v != 0) {
     if (index_type == phi::DataType::INT32) {
